refactor: move the Add impl for unsigned integers
